API Exception - TasksEventsDetails (Scheduled_Duration) data type change
git-svn-id: http://svn.gavstech.com/RAS-BAS/CSP/SourceCode/Main@6399 70b15045-743b-414c-a502-22bcac4871f6

+++ GAVS.AllocationSystem/GAVS.AllocationSystem.Model/CSP/SP/TasksEventsDetails.cs
@@ -18,7 +18,7 @@ namespace GAVS.AllocationSystem.Model.CSP
         public string description { get; set; }
         public string status { get; set; }
         public DateTime? scheduledStartDate { get; set; }
-        public int? scheduledDuration { get; set; }
+        public decimal? scheduledDuration { get; set; }
         public string priority { get; set; }
         public DateTime? dueDate { get; set; }
         public string owner { get; set; }
